fix(mookme): fix post-commit hook with previous commit strategy
